% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoutFusions.R
\name{ajoutFusions}
\alias{ajoutFusions}
\title{Procedure de traitement des fusions de communes definies  par RADE ou l'INSEE.}
\usage{
ajoutFusions(dataCommIGN_ZR_A, chgmtsINSEE, valZR, verbose = FALSE)
}
\arguments{
\item{dataCommIGN_ZR_A}{Data.frame. Contient les donnees Admin Express avec les informations sur les zonages}

\item{chgmtsINSEE}{Data.frame. Contient l'ensemble des changements INSEE sur les communes.}

\item{valZR}{Vecteur. Contient les differentes valeurs attendues de zonages.}

\item{verbose}{Booleen. Indique si les messages d'information sont affiches.}
}
\value{
Data.frame.
}
\description{
Procedure de traitement des fusions de communes definies  par RADE ou l'INSEE.
}
