% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoutInfosAgence.R
\name{ajoutInfosAgence}
\alias{ajoutInfosAgence}
\title{Procedure de lecture des changements sur les communes specifiees par l'Agence.}
\usage{
ajoutInfosAgence(dataCommIGN_ZR_A, fichierChgmtsAgence, valZR, verbose = FALSE)
}
\arguments{
\item{dataCommIGN_ZR_A}{Data.frame. Contient les donnees Admin Express avec les informations sur les zonages}

\item{fichierChgmtsAgence}{Chaine de caracteres. Chemin vers le fichier des changements de l'Agence.}

\item{valZR}{Vecteur. Contient les differentes valeurs attendues de zonages.}

\item{verbose}{Booleen. Indique si les messages d'information sont affiches.}
}
\value{
Data.frame
}
\description{
Procedure de lecture des changements sur les communes specifiees par l'Agence.
}
