% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoutScissions.R
\name{ajoutScissions}
\alias{ajoutScissions}
\title{Procedure de traitement des scissions de communes definies  par RADE ou l'INSEE ou par le fichier des changements specifique.}
\usage{
ajoutScissions(
  dataCommIGN_ZR_A,
  chgmtsINSEE,
  valZR,
  fichierChgmtsAgence = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dataCommIGN_ZR_A}{Data.frame. Contient les donnees Admin Express avec les informations sur les zonages}

\item{chgmtsINSEE}{Data.frame. Contient l'ensemble des changements INSEE sur les communes.}

\item{valZR}{Vecteur. Contient les differentes valeurs attendues de zonages.}

\item{fichierChgmtsAgence}{Chaine de caracteres. Chemin vers le fichier des changements specifiques.}

\item{verbose}{Booleen. Indique si les messages d'information sont affiches.}
}
\value{
Data.frame
}
\description{
Procedure de traitement des scissions de communes definies  par RADE ou l'INSEE ou par le fichier des changements specifique.
}
