% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creationZR.R
\name{creationZR}
\alias{creationZR}
\title{Fonction principale permettant de generer les couches des zonages de revedevances a partir des donnees de RADE ou de l'INSEE.}
\usage{
creationZR(
  annee_n,
  d_projetZR,
  code_bassin,
  nouveau_prog = FALSE,
  d_admin_express = NULL,
  d_ZRE = NULL,
  zre_esu = TRUE,
  zre_eso = TRUE,
  f_commzr_a = NULL,
  insee_source = "RADE",
  chgmtsAgence = FALSE,
  valZR = c("BASE", "MOYENNE", "RENFORCEE", "ZTQ"),
  ignoreCoherence = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{annee_n}{Entier. Represente l'annee etudiee, pour laquelle les zonages doivent etre crees}

\item{d_projetZR}{Chaine de caracteres. Chemin vers le dossier racine du projet. Attention: mettre des '/' pour le chemin, y compris sous Windows.}

\item{code_bassin}{Chaine de caracteres. Indique le code SANDRE de la circonscription de bassin souhaitee.}

\item{nouveau_prog}{Booleen. Indique si un nouveau programme debute a l'annee etudiee. La procedure se basera sur le fichier de communes de l'annee N et non celui de l'annee N-1.}

\item{d_admin_express}{Chaine de caracteres. Facultatif. Chemin vers le dossier contenant les donnees Admin Express structuree par annee. Attention: mettre des '/' pour le chemin, y compris sous Windows.}

\item{d_ZRE}{Chaine de caracteres. Facultatif. Chemin vers le dossier contant les couches ZRE structurees par type : ESU, ESO}

\item{zre_esu}{Booleen. Indique si les ZRE ESU doivent etre prises en compte.}

\item{zre_eso}{Booleen. Indique si les ZRE ESO doivent etre prises en compte.}

\item{f_commzr_a}{Chaine de caracteres. Facultatif. Chemin complet vers le fichier des zonages de l'annee precedente. Attention: mettre des '/' pour le chemin, y compris sous Windows.}

\item{insee_source}{Chaine de caracteres. Facultatif. Defini la source des donnees INSEE. RADE ou INSEE. RADE fait reference au projet de l'AESN. INSEE correspond aux fichiers a disposition sur le site de l'INSEE.}

\item{chgmtsAgence}{Booleen. Indique si des changements specifiques doivent etre pris en compte. Cela sert notament pour les changements de programme, de communes sortantes ou entrantes. Un fichier specifique est a remplir. Le script va creer un fichier vierge s'il n'existe pas.}

\item{valZR}{Vecteur. Contient les differentes valeurs attendues de zonages.}

\item{ignoreCoherence}{Booleen. Indique si la procedure de verification entre les codes INSEE du tableau auquel les fusions et scissions ont ete applique, correspond a la liste des codes INSEE attendus par les donnees INSEE.}

\item{verbose}{Booleen. Indique si les messages d'information sont affiches.}
}
\value{
un booleen. TRUE si le traitement a reussi.
}
\description{
Fonction principale permettant de generer les couches des zonages de revedevances a partir des donnees de RADE ou de l'INSEE.
}
