% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometriesZonages.R
\name{geometriesZonages}
\alias{geometriesZonages}
\title{Procedure d'export des couches SIG des zonages de redevances avec les decoupage sur les ZRE si necessaire.}
\usage{
geometriesZonages(
  annee_n,
  d_projetZR,
  commIGN_ZR_N,
  zre_esu = TRUE,
  zre_eso = TRUE,
  d_ZRE = NULL,
  verbose = FALSE
)
}
\arguments{
\item{annee_n}{Entier. Represente l'annee etudiee, pour laquelle les zonages doivent etre crees}

\item{d_projetZR}{Chaine de caracteres. Chemin vers le dossier racine du projet. Attention: mettre des '/' pour le chemin, y compris sous Windows.}

\item{commIGN_ZR_N}{Data.frame. Donnees contenant la liste des codes INSEE des communes et les zonages affectes a chaque commune}

\item{zre_esu}{Booleen. Indique si les zonages ZRE ESU sont pris en compte.}

\item{zre_eso}{Booleen. Indique si les zonages ZRE ESO sont pris en compte.}

\item{d_ZRE}{Chaine de caracteres. Facultatif. Chemin vers le dossier contant les couches ZRE structurees par type : ESU, ESO}

\item{verbose}{Booleen. Indique si les messages d'information sont affiches.}
}
\description{
Procedure d'export des couches SIG des zonages de redevances avec les decoupage sur les ZRE si necessaire.
}
