% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectureAdminExpress.R
\name{lectureAdminExpress}
\alias{lectureAdminExpress}
\title{Procedure de lecture des donnees Admin Express}
\usage{
lectureAdminExpress(annee, d_admin_express, carto = TRUE, verbose = FALSE)
}
\arguments{
\item{annee}{Entier. Annee du referentiel Admin Express}

\item{d_admin_express}{Chaine de caracteres. Chemin vers le dossier contenant les donnees Admin Express structuree par annee. Attention: mettre des '/' pour le chemin, y compris sous Windows.}

\item{carto}{Booleen. Indique si la fonction privilegie la couche "_CARTO" de l'ADmin Express = version simplifiee des contours des communes}

\item{verbose}{Booleen. Indique si les messages d'information sont affiches.}
}
\value{
SpatialPolygonsDataFrame.
}
\description{
Procedure de lecture des donnees Admin Express
}
