% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectureINSEE.R
\name{lectureINSEE}
\alias{lectureINSEE}
\title{Procedure de lecture des donnees INSEE provenant de RADE ou des fichiers de l'INSEE}
\usage{
lectureINSEE(
  annee_n,
  code_bassin,
  d_projetZR = NULL,
  type = "RADE",
  verbose = FALSE
)
}
\arguments{
\item{annee_n}{Entier. Annee du referentiel INSEE}

\item{code_bassin}{Chaine de caracteres. Indique le code SANDRE de la circonscription de bassin souhaitee.}

\item{d_projetZR}{Chaine de caracteres. Chemin vers le dossier racine du projet. Attention: mettre des '/' pour le chemin, y compris sous Windows.}

\item{type}{Chaine de caracteres. Indique la source de la donnee. RADE ou INSEE}

\item{verbose}{Booleen. Indique si les messages d'information sont affiches.}
}
\value{
Data.frame
}
\description{
Procedure de lecture des donnees INSEE provenant de RADE ou des fichiers de l'INSEE
}
