% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_ZR_Fusion.R
\name{selection_ZR_Fusion}
\alias{selection_ZR_Fusion}
\title{Fonction permettant de definir le bon zonage de redevance lors d'une fusion de plusieurs commune. Regle : Zonage final = Plus grane surface par zonage.}
\usage{
selection_ZR_Fusion(dataCommIGN_ZR_A, champ_zr, inseeMeres, verbose = FALSE)
}
\arguments{
\item{dataCommIGN_ZR_A}{Data.frame. Contient les donnees Admin Express avec les informations sur les zonages}

\item{champ_zr}{Chaine de caracteres. Indique le champ (= type de zonage) etudie.}

\item{inseeMeres}{Vecteur. Contient les codes INSEE des communes qui fusionnent}

\item{verbose}{Booleen. Indique si les messages d'information sont affiches.}
}
\value{
Chaine de caracteres.
}
\description{
Fonction permettant de definir le bon zonage de redevance lors d'une fusion de plusieurs commune. Regle : Zonage final = Plus grane surface par zonage.
}
