% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppArtefact.R
\name{suppArtefact}
\alias{suppArtefact}
\title{Fonction de suppression des polygons artefacts suite a la découpe des ZRE.}
\usage{
suppArtefact(poly, limArea)
}
\arguments{
\item{poly}{SpatialPolygonsDataFrame. Couche sur laquelle les polygones inferieur à la surface limite seront supprimes.}

\item{limArea}{Entier. Surface limite des polygones. En dessous de cette limite, tous les polygones sont supprimes.}
}
\value{
SpatialPolygonsDataFrame.
}
\description{
Fonction de suppression des polygons artefacts suite a la découpe des ZRE.
}
